Added include for stddef.h cause test needs it and doesn't include it.

+++ linked_list.h
@@ -2,6 +2,7 @@
 #define LINKED_LIST_H
 #include <stdint.h>
 #include <stdlib.h>
+#include <stddef.h>
 
 #include "common_defs.h"
 #include "memory_manager.h"
@@ -13,15 +14,15 @@ typedef struct __attribute__((packed)) Node {
 
 void list_init(Node** head, size_t size);
 
-void list_insert(Node** head, int data);
+void list_insert(Node** head, uint16_t data);
 
-void list_insert_after(Node* prev_node, int data);
+void list_insert_after(Node* prev_node, uint16_t data);
 
-void list_insert_before(Node** head, Node* next_node, int data);
+void list_insert_before(Node** head, Node* next_node, uint16_t data);
 
-void list_delete(Node** head, int data);
+void list_delete(Node** head, uint16_t data);
 
-Node* list_search(Node** head, int data);
+Node* list_search(Node** head, uint16_t data);
 
 void list_display(Node** head);
